pyproject.toml
[project]
name = "digichecks-shacl-service"
version = "1.0.0"
description = "SHACL Compliancy Check Service developed for the DigiChecks Horizon Project"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "GPL-3.0-or-later"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "cryptography==43.0.1",
    "fastapi==0.110.0",
    "gunicorn==23.0.0",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "passlib==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic-settings==2.2.1",
    "pyshacl==0.26.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "rdflib==7.0.0",
    "requests==2.32.3",
    "uvicorn==0.27.1",
    "validators==0.34.0",
]
